added JpaWordRepositoryTest

+++ src/main/java/com/maximKachan/englishDictionary/repository/WordRepository.java
@@ -8,11 +8,11 @@ import java.util.List;
 public interface WordRepository {
     List<Word> getWords(String pattern) throws DaoException;
 
-    Word getWordById(Long id) throws DaoException;
+    Word getWordById(Integer id) throws DaoException;
 
     void addWord(Word word) throws DaoException;
 
-    void updateWord(Long id, Word word) throws DaoException;
+    void updateWord(Integer id, Word word) throws DaoException;
 
-    void deleteWord(Long id) throws DaoException;
+    void deleteWord(Integer id) throws DaoException;
 }
